// FrontEnd/runbook-app/src/model/alert.ts
export class Alert{

    constructor(init?: Partial<Alert>){
        Object.assign(this,init);
    }
    _id: String = "";
    col1: string = "";
    col2: string = "";
    col3: string = "";
    col4: string = "";
    col5: string= "";
    col6: string= "";
    col7: string= "";
    col8: string= "";
    col9: string= "";
    col10: string= "";
    col11: string= "";
    col12: string= "";
    col13: string= "";
    col14: string= "";
    col15: string= "";
    col16: string= "";
    col17: string= "";
    col18: string= "";
    col19: string= "";
    col20: string= "";
    col21: string = "";
};